USGS-Earthquakes/airflow/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
dlt[snowflake]>=1.22.2
dbt-snowflake==1.11.2
astronomer-cosmos
apache-airflow-providers-snowflake